ParticleEngine structure done and tested
Infrastructure for particle engine is done, need to add some more
particle effects and textures.

+++ MobileGame/MobileGame/GameObjects/Towers/TowerLight.cs
@@ -4,7 +4,7 @@ namespace MobileGame.GameObjects.Towers
 {
 	public class TowerLight: Tower
 	{
-		public TowerLight(float range, float dmg, float timeBetweenShoots, RenderDesc renderDesc)
+		public TowerLight(float range, int dmg, float timeBetweenShoots, RenderDesc renderDesc)
 			: base(range, dmg, timeBetweenShoots, renderDesc)
 		{
 		}
